\name{OrderedBart-package}
\alias{OrderedBart-package}
\alias{OrderedBart}
\docType{package}
\title{
\packageTitle{OrderedBart}
}
\description{
\packageDescription{OrderedBart}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OrderedBart}
\packageIndices{OrderedBart}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{OrderedBart}

Maintainer: \packageMaintainer{OrderedBart}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
